Integrate wavelet reporting with thresholding module
